Split SignerSuite into Java (in)dependent parts
