+++ core/shared/src/test/scala/bobcats/util/PEMUtils.scala
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2021 Typelevel
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package bobcats.util
+
+import bobcats.{PrivateKeySpec, PublicKeySpec}
+
+import scala.util.Try
+
+trait PEMUtils {
+	//todo
+	//def pemToKey(pemStr: String): Try[Either[PrivateKeySpec[_],PublicKeySpec[_]]]
+	def getPrivateKeyFromPEM(pemStr: String): Try[PrivateKeySpec[_]]
+	def getPublicKeyFromPEM(pemStr: String): Try[PublicKeySpec[_]]
+}
